+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [1.11.0] - Unreleased
+## [1.11.0] - 2022-09-01
 ### Added
 - New plugin `minify_html` [#248].
 - New plugin `multilanguage`, to create pages of different languages [#205].
@@ -1734,7 +1734,7 @@ The first version.
 [#249]: https://github.com/lumeland/lume/issues/249
 [#250]: https://github.com/lumeland/lume/issues/250
 
-[1.11.0]: https://github.com/lumeland/lume/compare/v1.10.4...HEAD
+[1.11.0]: https://github.com/lumeland/lume/compare/v1.10.4...v1.11.0
 [1.10.4]: https://github.com/lumeland/lume/compare/v1.10.3...v1.10.4
 [1.10.3]: https://github.com/lumeland/lume/compare/v1.10.2...v1.10.3
 [1.10.2]: https://github.com/lumeland/lume/compare/v1.10.1...v1.10.2